+++ CHANGELOG.md
@@ -75,6 +75,7 @@ This release contains several breaking changes to 1.6 due to heavy internal refa
 - When switching directories while a search result list is displayed, this search is now performed at the other directory immediately after switching.
 - Reversing a MagicQuote can now be performed by pressing backspace after a replacement has taken place, in order to restore the default double (") or single (') quote.
 - Math doesn't render in comments anymore.
+- Opening files with Zettlr when the app is not running will now correctly open them.
 
 ## Under the Hood